pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = ""
authors = ["Izanderson Florencio <izandersonf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
mysqlclient = "^2.1.0"
ipython = "^8.2.0"
Django = "^4.0.4"
python-decouple = "^3.6"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
